+++ src/tracing/tracingUtil.js
@@ -23,6 +23,11 @@ exports.getStackTrace = function getStackTrace(referenceFunction) {
   return stackTrace.captureStackTrace(stackTraceLength, referenceFunction);
 };
 
+exports.generateRandomTraceId = function generateRandomTraceId() {
+  // TODO: as soon as all Instana tracers support 128bit trace IDs we can
+  // turn this into generating a longer String
+  return Math.floor(Math.random() * Number.MAX_SAFE_INTEGER).toString(16);
+};
 
 exports.generateRandomSpanId = function generateRandomSpanId() {
   return Math.floor(Math.random() * Number.MAX_SAFE_INTEGER).toString(16);